FL-1345: Support for QA builds.
If $[release/type] is set to "official", we use the existing production-oriented
build approach.

If $[release/type] is anything else, we do QA builds. For QA builds, typically we
want to use an existing production stage3 as a seed. QA builds don't build stages
of their own (this may change). funtoo-current-next is an example QA build.

For QA builds, failcount isn't touched. A status file is now written to the build
directory containing "ok" for build success or "fail" for build failure.

Current use right now is to set up a QA build with a build strategy of remote,
and specify the SHA1 of the commit in the portage tree to use for testing. The
QA build will be in a "<SHA1>" dir instead of a "2015-04-03" dir.