Remove all creation of device nodes from metro.
This should allow metro to run inside an unprivileged container. But,
we need to test this. And potentially need an additional script to run
on stage3 install on bare metal to create necessary set of minimal
device nodes at boot.

So these changes should be considered EXPERIMENTAL. But we will be
actively testing them and getting them working as we want metro to
work without issue in an unprivileged lxd container.