This fixes some concepts related to digest generation.
We used to have each build box be an island. Now the model is to
have them share a build repository using NFS or another network
filesystem. So we can't generate GPG signatures for anything we
find -- we should just sign what we actually built locally.
These fixes added to buildrepo digestgen.

Also added zero-length .cme files to mark an associated tarball
as needing compression. This eliminates race condition where we
may accidentally remotely try to compress a tarball that is still
being generated. So it should be safe for a remote compression
agent to scan for .cme files and just compress these files.

Also added a buildrepo compress which will do a global scan of
the build repo and compress any non-compressed files it finds,
even non-locally-built files. So you can use a fast server to
compress artifacts.